src/test/javascript/spec/app/entities/combo/combo-update.component.spec.ts
/* tslint:disable max-line-length */
import { shallowMount, createLocalVue, Wrapper } from '@vue/test-utils';
import sinon, { SinonStubbedInstance } from 'sinon';
import Router from 'vue-router';
import { ToastPlugin } from 'bootstrap-vue';

import * as config from '@/shared/config/config';
import ComboUpdateComponent from '@/entities/combo/combo-update.vue';
import ComboClass from '@/entities/combo/combo-update.component';
import ComboService from '@/entities/combo/combo.service';

import AlertService from '@/shared/alert/alert.service';

const localVue = createLocalVue();

config.initVueApp(localVue);
const i18n = config.initI18N(localVue);
const store = config.initVueXStore(localVue);
const router = new Router();
localVue.use(Router);
localVue.use(ToastPlugin);
localVue.component('font-awesome-icon', {});
localVue.component('b-input-group', {});
localVue.component('b-input-group-prepend', {});
localVue.component('b-form-datepicker', {});
localVue.component('b-form-input', {});

describe('Component Tests', () => {
  describe('Combo Management Update Component', () => {
    let wrapper: Wrapper<ComboClass>;
    let comp: ComboClass;
    let comboServiceStub: SinonStubbedInstance<ComboService>;

    beforeEach(() => {
      comboServiceStub = sinon.createStubInstance<ComboService>(ComboService);

      wrapper = shallowMount<ComboClass>(ComboUpdateComponent, {
        store,
        i18n,
        localVue,
        router,
        provide: {
          comboService: () => comboServiceStub,
          alertService: () => new AlertService(),
        },
      });
      comp = wrapper.vm;
    });

    describe('save', () => {
      it('Should call update service on save for existing entity', async () => {
        // GIVEN
        const entity = { id: 123 };
        comp.combo = entity;
        comboServiceStub.update.resolves(entity);

        // WHEN
        comp.save();
        await comp.$nextTick();

        // THEN
        expect(comboServiceStub.update.calledWith(entity)).toBeTruthy();
        expect(comp.isSaving).toEqual(false);
      });

      it('Should call create service on save for new entity', async () => {
        // GIVEN
        const entity = {};
        comp.combo = entity;
        comboServiceStub.create.resolves(entity);

        // WHEN
        comp.save();
        await comp.$nextTick();

        // THEN
        expect(comboServiceStub.create.calledWith(entity)).toBeTruthy();
        expect(comp.isSaving).toEqual(false);
      });
    });

    describe('Before route enter', () => {
      it('Should retrieve data', async () => {
        // GIVEN
        const foundCombo = { id: 123 };
        comboServiceStub.find.resolves(foundCombo);
        comboServiceStub.retrieve.resolves([foundCombo]);

        // WHEN
        comp.beforeRouteEnter({ params: { comboId: 123 } }, null, cb => cb(comp));
        await comp.$nextTick();

        // THEN
        expect(comp.combo).toBe(foundCombo);
      });
    });

    describe('Previous state', () => {
      it('Should go previous state', async () => {
        comp.previousState();
        await comp.$nextTick();

        expect(comp.$router.currentRoute.fullPath).toContain('/');
      });
    });
  });
});
